requirements.txt
openai>=1.1.1
streamlit
python-dotenv
